ROBUSTNESS: checkDir() does a better/more robust job on Windows [#65]
It seems as if `file.exists()` can return `FALSE` while the same file is
still seen by `list.files()`.  This causes some of the sanity checks of
BatchJobs to fail, which is exactly what I reported in #65.

Because of this, I have now replaced `unlink()` with an internal
`removeDirs()` which calls `unlink()` multiple times if needed and it not
only checks with `file.exists()` whether the file/directory is removed,
but also with `list.files()`.  If still there, it sleeps a bit and
retries.  It does this many time before giving up.

I also robustified the tests for creating files done by `is.accessible()`.
It not only use `file.create()`, but also writes to the file created and
makes sure the content can be read back properly. If not, an error is
thrown.

I have now run the torture tests on Windows 7 64-bit (NTFS) creating
and removing 3500 registries over and over quickly without errors.
Previously, it failed every 10-50 registries created the same way.